Also move HTTPClient to bravado
