+++ bravado_core/content_type.py
@@ -1,0 +1,1 @@
+APP_JSON = 'application/json'
